+++ src/client/components/evaluate/detail/MajorityRace.tsx
@@ -1,10 +1,17 @@
 /** @jsx jsx */
-import { Box, Flex, jsx, Styled, ThemeUIStyleObject, Heading } from "theme-ui";
-import { DistrictsGeoJSON, EvaluateMetricWithValue } from "../../../types";
-import Tooltip from "../../Tooltip";
-import DemographicsTooltip from "../../DemographicsTooltip";
-import DemographicsChart from "../../DemographicsChart";
+import { connect } from "react-redux";
+import { Box, Flex, Heading, jsx, Styled, ThemeUIStyleObject } from "theme-ui";
+
+import { GroupTotal, IStaticMetadata } from "../../../../shared/entities";
+
 import { getMajorityRaceDisplay, isMajorityMinority } from "../../../functions";
+import { State } from "../../../reducers";
+import { DistrictsGeoJSON, EvaluateMetricWithValue } from "../../../types";
+
+import DemographicsChart from "../../DemographicsChart";
+import DemographicsTooltip from "../../DemographicsTooltip";
+import Tooltip from "../../Tooltip";
+import { getDemographicsGroups } from "../../../../shared/functions";
 
 const style: ThemeUIStyleObject = {
   table: {
@@ -60,13 +67,21 @@ const style: ThemeUIStyleObject = {
   }
 };
 
+interface StateProps {
+  readonly populationKey: GroupTotal;
+}
+
 const MajorityRaceMetricDetail = ({
   metric,
-  geojson
+  geojson,
+  metadata,
+  populationKey
 }: {
   readonly metric: EvaluateMetricWithValue;
+  readonly metadata?: IStaticMetadata;
   readonly geojson?: DistrictsGeoJSON;
-}) => {
+} & StateProps) => {
+  const demographicsGroups = metadata ? getDemographicsGroups(metadata) : [];
   return (
     <Box>
       <Heading as="h2" sx={{ variant: "text.h5", mt: 4 }}>
@@ -81,62 +96,75 @@ const MajorityRaceMetricDetail = ({
           </Styled.tr>
         </thead>
         <tbody>
-          {geojson?.features.map(
-            (feature, id) =>
-              id > 0 && (
-                <Styled.tr key={id}>
-                  <Styled.td sx={{ ...style.td, ...style.colFirst }}>{id}</Styled.td>
+          {metadata &&
+            geojson?.features.map(
+              (feature, id) =>
+                id > 0 && (
+                  <Styled.tr key={id}>
+                    <Styled.td sx={{ ...style.td, ...style.colFirst }}>{id}</Styled.td>
 
-                  <Styled.td sx={style.td}>
-                    {feature.properties.majorityRace !== undefined ? (
-                      <Flex sx={{ alignItems: "center" }}>
-                        <Styled.div
-                          sx={{
-                            mr: 2,
-                            width: "15px",
-                            height: "15px",
-                            borderRadius: "small",
-                            bg: feature.properties.majorityRaceFill
-                          }}
-                        ></Styled.div>
-                        <Box>{getMajorityRaceDisplay(feature)}</Box>
-                      </Flex>
-                    ) : (
-                      <Box sx={style.blankValue}>-</Box>
-                    )}
-                  </Styled.td>
+                    <Styled.td sx={style.td}>
+                      {feature.properties.majorityRace !== undefined ? (
+                        <Flex sx={{ alignItems: "center" }}>
+                          <Styled.div
+                            sx={{
+                              mr: 2,
+                              width: "15px",
+                              height: "15px",
+                              borderRadius: "small",
+                              bg: feature.properties.majorityRaceFill
+                            }}
+                          ></Styled.div>
+                          <Box>{getMajorityRaceDisplay(feature)}</Box>
+                        </Flex>
+                      ) : (
+                        <Box sx={style.blankValue}>-</Box>
+                      )}
+                    </Styled.td>
 
-                  <Styled.td sx={{ ...style.td, ...style.colLast }}>
-                    <Tooltip
-                      placement="top-start"
-                      content={
-                        feature.properties.demographics.population !== 0 ? (
-                          <DemographicsTooltip
-                            demographics={feature.properties.demographics}
-                            isMajorityMinority={isMajorityMinority(feature)}
-                          />
-                        ) : (
-                          <em>
-                            <strong>Empty district.</strong> Add people to this district to view the
-                            race chart
-                          </em>
-                        )
-                      }
-                    >
-                      <span sx={{ display: "inline-block" }}>
-                        <span sx={style.chart}>
-                          <DemographicsChart demographics={feature.properties.demographics} />
+                    <Styled.td sx={{ ...style.td, ...style.colLast }}>
+                      <Tooltip
+                        placement="top-start"
+                        content={
+                          feature.properties.demographics.population !== 0 ? (
+                            <DemographicsTooltip
+                              demographics={feature.properties.demographics}
+                              isMajorityMinority={isMajorityMinority(feature)}
+                              demographicsGroups={demographicsGroups}
+                              populationKey={populationKey}
+                            />
+                          ) : (
+                            <em>
+                              <strong>Empty district.</strong> Add people to this district to view
+                              the race chart
+                            </em>
+                          )
+                        }
+                      >
+                        <span sx={{ display: "inline-block" }}>
+                          <span sx={style.chart}>
+                            <DemographicsChart
+                              demographics={feature.properties.demographics}
+                              demographicsGroups={demographicsGroups}
+                              populationKey={populationKey}
+                            />
+                          </span>
                         </span>
-                      </span>
-                    </Tooltip>
-                  </Styled.td>
-                </Styled.tr>
-              )
-          )}
+                      </Tooltip>
+                    </Styled.td>
+                  </Styled.tr>
+                )
+            )}
         </tbody>
       </Styled.table>
     </Box>
   );
 };
 
-export default MajorityRaceMetricDetail;
+function mapStateToProps(state: State): StateProps {
+  return {
+    populationKey: state.projectOptions.populationKey
+  };
+}
+
+export default connect(mapStateToProps)(MajorityRaceMetricDetail);