+++ src/client/reducers/districtDrawing.ts
@@ -2,7 +2,7 @@ import { Cmd, Loop, loop, LoopReducer } from "redux-loop";
 import { getType } from "typesafe-actions";
 
 import { Action } from "../actions";
-import { SavingState, EvaluateMetricWithValue, ElectionYear } from "../types";
+import { SavingState, EvaluateMetricWithValue } from "../types";
 
 import {
   addSelectedGeounits,
@@ -17,10 +17,6 @@ import {
   setSelectedDistrictId,
   setHoveredDistrictId,
   setSelectionTool,
-  showAdvancedEditingModal,
-  showCopyMapModal,
-  showConvertMapModal,
-  setImportFlagsModal,
   toggleDistrictLocked,
   undo,
   replaceSelectedGeounits,
@@ -31,18 +27,16 @@ import {
   saveDistrictsDefinition,
   setSavingState,
   FindTool,
-  toggleLimitDrawingToWithinCounty,
   selectEvaluationMetric,
   setZoomToDistrictId,
   setMapLabel,
-  toggleKeyboardShortcutsModal,
-  setElectionYear,
   PaintBrushSize,
   setPaintBrushSize,
   toggleExpandedMetrics,
   toggleReferenceLayer
 } from "../actions/districtDrawing";
 import { updateDistrictsDefinition, updateDistrictLocks } from "../actions/projectData";
+import { showAdvancedEditingModal } from "../actions/projectModals";
 import { SelectionTool } from "../actions/districtDrawing";
 import { resetProjectState } from "../actions/root";
 import {
@@ -119,19 +113,12 @@ export interface DistrictDrawingState {
   readonly highlightedGeounits: GeoUnits;
   readonly selectionTool: SelectionTool;
   readonly paintBrushSize: PaintBrushSize;
-  readonly showAdvancedEditingModal: boolean;
-  readonly showCopyMapModal: boolean;
-  readonly showConvertMapModal: boolean;
-  readonly showKeyboardShortcutsModal: boolean;
-  readonly showImportFlagsModal: boolean;
   readonly findMenuOpen: boolean;
   readonly expandedProjectMetrics: boolean;
   readonly evaluateMode: boolean;
   readonly evaluateMetric: EvaluateMetricWithValue | undefined;
   readonly findIndex?: number;
   readonly findTool: FindTool;
-  readonly limitSelectionToCounty: boolean;
-  readonly electionYear: ElectionYear;
   readonly saving: SavingState;
   readonly undoHistory: UndoHistory;
   readonly mapLabel: string | undefined;
@@ -145,18 +132,11 @@ export const initialDistrictDrawingState: DistrictDrawingState = {
   highlightedGeounits: {},
   selectionTool: SelectionTool.Default,
   paintBrushSize: 1,
-  showAdvancedEditingModal: false,
-  showCopyMapModal: false,
-  showConvertMapModal: false,
-  showImportFlagsModal: false,
-  showKeyboardShortcutsModal: false,
   findMenuOpen: false,
   expandedProjectMetrics: false,
   evaluateMode: false,
   evaluateMetric: undefined,
   findTool: FindTool.Unassigned,
-  limitSelectionToCounty: false,
-  electionYear: "16",
   saving: "unsaved",
   mapLabel: "undefined",
   showReferenceLayers: new Set(),
@@ -256,16 +236,6 @@ const districtDrawingReducer: LoopReducer<ProjectState, Action> = (
         ...state,
         saving: action.payload
       };
-    case getType(toggleLimitDrawingToWithinCounty):
-      return {
-        ...state,
-        limitSelectionToCounty: !state.limitSelectionToCounty
-      };
-    case getType(setElectionYear):
-      return {
-        ...state,
-        electionYear: action.payload
-      };
     case getType(setHighlightedGeounits):
       return {
         ...state,
@@ -330,31 +300,6 @@ const districtDrawingReducer: LoopReducer<ProjectState, Action> = (
           )
         )
       );
-    case getType(showAdvancedEditingModal):
-      return {
-        ...state,
-        showAdvancedEditingModal: action.payload
-      };
-    case getType(toggleKeyboardShortcutsModal):
-      return {
-        ...state,
-        showKeyboardShortcutsModal: !state.showKeyboardShortcutsModal
-      };
-    case getType(showCopyMapModal):
-      return {
-        ...state,
-        showCopyMapModal: action.payload
-      };
-    case getType(showConvertMapModal):
-      return {
-        ...state,
-        showConvertMapModal: action.payload
-      };
-    case getType(setImportFlagsModal):
-      return {
-        ...state,
-        showImportFlagsModal: action.payload
-      };
     case getType(toggleFind):
       return {
         ...state,